Separated tests for LastUpdated et LastUpdatedConnected
