+++ src/components/lastUpdated/tests/LastUpdated.spec.tsx
@@ -1,16 +1,11 @@
 import { shallow, mount, ReactWrapper, ShallowWrapper } from 'enzyme';
 import { LastUpdated, ILastUpdatedProps, defaultLastUpdateLabel } from '../LastUpdated';
-import { IReactVaporState, clearState } from '../../../utils/ReduxUtils';
-import { LastUpdatedConnected } from '../LastUpdatedConnected';
 import { TestUtils } from '../../../utils/TestUtils';
-import { Store } from 'redux';
-import { Provider } from 'react-redux';
 import * as s from 'underscore.string';
 import * as moment from 'moment';
 
 /* tslint:disable:no-unused-variable */
 import * as React from 'react';
-import { changeLastUpdated } from '../LastUpdatedActions';
 /* tslint:enable:no-unused-variable */
 
 describe('LastUpdated', () => {
@@ -112,79 +107,6 @@ describe('LastUpdated', () => {
       lastUpdated.mount();
       lastUpdated.unmount();
       expect(destroySpy.calls.count()).toBe(1);
-    });
-  });
-
-  describe('<LastUpdatedConnected />', () => {
-    let wrapper: ReactWrapper<any, any>;
-    let store: Store<IReactVaporState>;
-
-    beforeEach(() => {
-      id = 'last-update';
-
-      store = TestUtils.buildStore();
-
-      wrapper = mount(
-        <Provider store={store}>
-          <LastUpdatedConnected
-            id={id}
-            />
-        </Provider>,
-        { attachTo: document.getElementById('App') }
-      );
-      lastUpdated = wrapper.find(LastUpdated).first();
-    });
-
-    afterEach(() => {
-      store.dispatch(clearState());
-      wrapper.unmount();
-      wrapper.detach();
-    });
-
-    it('should get its id as a prop', () => {
-      let idProp = lastUpdated.props().id;
-
-      expect(idProp).toBeDefined();
-      expect(idProp).toBe(id);
-    });
-
-    it('should get last update time as a prop', () => {
-      let timeProp = lastUpdated.props().time;
-
-      expect(timeProp).toBeDefined();
-    });
-
-    it('should get what to do on render as a prop', () => {
-      let onRenderProp = lastUpdated.props().onRender;
-
-      expect(onRenderProp).toBeDefined();
-    });
-
-    it('should get what to do on destroy as a prop', () => {
-      let onDestroyProp = lastUpdated.props().onDestroy;
-
-      expect(onDestroyProp).toBeDefined();
-    });
-
-    it('should display the last update time', () => {
-      expect(s.contains(lastUpdated.html(), 'AM') || s.contains(lastUpdated.html(), 'PM')).toBe(true);
-    });
-
-    it('should add the last update time in the store on render', () => {
-      expect(store.getState().lastUpdatedComposite.filter(timer => timer.id === id).length).toBe(1);
-    });
-
-    it('should update the last update time in the store when dispatching a "changeLastUpdated" action', () => {
-      expect(store.getState().lastUpdatedComposite.filter(timer => timer.id === id).length).toBe(1);
-
-      let storedTime = store.getState().lastUpdatedComposite.filter(timer => timer.id === id)[0].time;
-      store.dispatch(changeLastUpdated(id));
-      expect(store.getState().lastUpdatedComposite.filter(timer => timer.id === id)[0]).not.toBe(storedTime);
-    });
-
-    it('should remove the last update time in the store on destroy', () => {
-      wrapper.unmount();
-      expect(store.getState().lastUpdatedComposite.filter(timer => timer.id === id).length).toBe(0);
     });
   });
 });
